requirements.txt
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langgraph>=0.1.0
langchain-openai>=0.1.0
python-dotenv>=1.0.1
aiohttp>=3.9.5
pydantic>=2.7.0
beautifulsoup4>=4.12.3
